Nsight Systems analysis from an exported JSON events file
- getStats.py: reads a JSON CUDA events file from a JSON file that was
exported from an .nsys-rep file in nsys-ui. It prints a summary of:
   - global runtime statistics
   - prediced global speedups for a few typical local GPU speedups,
accoording to Amdahl's Law. This is only useful as a crude indicator.
   - Prints a Top-10 list of kernels (by kernel ID) sorted by total
execution time (get the most consuming kernel) and sorted by average
execution time (get the kernel that takes the longest to execute)
  - Plots a graph that shows the distribution of kernel execution times
on a log-log scale.

- getCorrelationID.sh:
   - with a kernel ID and without a stream number, prints a list of CUDA
streams where the kernel is launched
   - with a kernel ID and a stream number: lists the first correlationID
that can be used to find the kernels name in Nsight Systems CUDA Event
Viewer.